Initial version.
Parsing of C++ sources is implemented. During parse a call to the
"make_transition_table()" is detected (with or without namespace
"boost::sml") and its parameters are used as FSM transition table.

The initial states (marked with star) are defined as separate states
and transitions to them from UML initial state are deduplicated.

The program can accept a single parameter with file name to parse. If
not parameter is given the standard input is parsed.

The Mermaid diagram is generated to standard output.